requirements.txt
Flask
mysql-connector-python
flask-cors
gunicorn